requirements.txt
Django~=4.1.5
exif
pathlib
Pillow
pandas
django-colorfield